# [Operator] optimize normalize op with vectorized load, dynamic shape and more