this change introduces several enhancements to the current norm operator

* vectorized load for fp16 types
* allow epilogue
* dynamic shape on the normalized dimension 
* add a tuning to use 2 warp shuffle routines or just a single one
* cleaner code and implementation

as a result, norm_fp16.py can be safely deleted

for shapes in stable diffusion: [2, 32, 60, 16, 16], norm dims [60, 16, 16], fp32:
* torch : 0.023 ms
* main: 0.036 ms
* this change: 0.027 ms


for shapes in bert-base: [1, 128, 768], norm dims [768], fp16:
* torch: 0.006 ms
* main: 0.008 ms 
* this change: 0.007 ms

We are still not faster than torch, but very close